fix(LottoService): 수익률이 int 범위를 초과하는 경우에 대해 예외 처리

+++ src/main/java/lotto/domain/LottoService.java
@@ -25,7 +25,7 @@ public class LottoService {
         return ratingInfo;
     }
 
-    public double calculateEarningRate(final RatingInfo ratingInfo, final int money) {
+    public double calculateEarningRate(final RatingInfo ratingInfo, final long money) {
         return totalSum(ratingInfo) / money;
     }
